--- service-front/app/src/Common/src/Handler/WorkflowStep.php
@@ -1,14 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Common\Handler;
-
-interface WorkflowStep
-{
-    public function isMissingPrerequisite(): bool;
-
-    public function nextPage(): string;
-
-    public function lastPage(): string;
-}
